requirements.txt
Authlib==1.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
deprecation==2.1.0
Flask==2.3.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
python-dotenv==1.0.0
requests==2.30.0
urllib3==2.0.2
Werkzeug==2.3.4